build.rs: pass target-feature values directly to the compiler
The target-feature in the .cargo/config file generated by
the build.rs script are ignored on the first build.rs since
cargo does not reload the config file. They are applied
from the second build.
Passing those values directly to the compiler is better
since they are immediately used, also in the first build.