Improve rgb to yuv full-range performance
Conversion weights are adjusted to avoid integer overflow when converting
the color space.

In particular, the weights for rgb to u/v conversion are ensured not to
be greater than '32767' and symmetry is exploited, as for u/v wr+wg+wb should
sum to zero. For this reason, wb is simply evaluated to -(wr+wg).

Also, from the derivation it follows that the wr for v channel should
be equal to wb for u channel. Also such rule is applied in the weights
computation.

In the end, this allows to make such conversions functionally equivalent
to the narrow-range ones.